Level Load & Tooltip Interact
# Description
Added a script that asynchronously loads a given scene.

Added interacting to hovered objects by adding an additional method to the `ILookHoverer` interface.

The scripts that handles updating the tooltip now exposes to the editor an unity event to bind callbacks to when the tooltip is interacted with.

Fixed bug with `TooltipUI` not unbinding on disabled/destructed.