Scene Transition & Input
# Description
Added a script to receive an event callback from the animator events. The script notifies of transition start and end.

Added a scriptable object script for hashing parameter string names and verifying that an animator contains the parameter.

Added an object to the UI for the transition animation.

Added methods to request disabling and enabling gameplay inputs. Used to disable inputs on transition start.

Fixed `AssertReference` extension to output the class type instead of just "T". Ups.