DustBowl
# Description
The soil tiles are now a `FarmlandTile` that are infected by the `FarmlandManager`. The farmland manager notifies of the events related to the state of the tiles.

Added a new layer for interactables to avoid collision with the player.

Added material variants for the imported assets to reduce their reflection values that were too high.

Added a generic ui class that updates a text component to display a counter value. Used to display the count of affected tiles and withered tiles.

Added a terrain on the `dustbowl` level for decoration purposes.

Reorganised the prefabs folder.